+++ aws-runtime/http-client-engine-crt/common/src/aws/sdk/kotlin/runtime/http/engine/crt/SdkStreamResponseHandler.kt
@@ -169,6 +169,7 @@ internal class SdkStreamResponseHandler(
 
         if (forceClose) {
             conn.shutdown()
+            conn.close()
         }
     }
 }