+++ codegen/smithy-aws-kotlin-codegen/src/main/kotlin/aws/sdk/kotlin/codegen/AwsServiceConfigIntegration.kt
@@ -47,17 +47,18 @@ class AwsServiceConfigIntegration : KotlinIntegration {
         val serviceSymbol: Symbol = writer.getContextValue(ServiceGenerator.SectionServiceCompanionObject.ServiceSymbol)
         writer.withBlock("companion object {", "}") {
             withBlock(
-                "operator fun invoke(sharedConfig: #T? = null, block: Config.DslBuilder.() -> Unit = {}): #L {",
+                "operator fun invoke(sharedConfig: #T? = null, block: Config.Builder.() -> Unit = {}): #L {",
                 "}",
                 AwsRuntimeTypes.Types.AwsClientConfig,
                 serviceSymbol.name
             ) {
                 withBlock(
-                    "val config = Config.BuilderImpl().apply { ",
+                    "val config = Config.Builder().apply { ",
                     "}.apply(block).build()"
                 ) {
                     write("region = sharedConfig?.region")
                     write("credentialsProvider = sharedConfig?.credentialsProvider")
+                    write("sdkLogMode = sharedConfig?.sdkLogMode ?: SdkLogMode.Default")
                 }
                 write("return Default${serviceSymbol.name}(config)")
             }